\name{bcont}
\alias{bcont}
\alias{bcont3}
\alias{bcont23}
\alias{bcont32}
\alias{bcontROB}
\alias{ggm.Deriv}

\title{Internal Function}

\description{This and other similar internal functions provide the log-likelihood, gradient and observed information matrix for 
penalized/unpenalized maximum likelihood optimization when copula models with continuous margins are employed.

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}




