\name{bdiscrcont}
\alias{bdiscrcont}
\alias{bdiscrcont12}
\alias{bdiscrcont13}
\alias{bdiscrcont23}


\title{Internal Function}

\description{This and other similar internal functions provide the log-likelihood, gradient and observed information matrix for 
penalized/unpenalized maximum likelihood optimization when copula models with discrete and continuous margins are employed.

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}




