\name{bprobgHs}
\alias{bprobgHs}
\alias{bprobgHstwoParC}
\alias{bprobgHsBinROY}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and observed/Fisher information matrix for 
penalized/unpenalized maximum likelihood optimization when copula models with binary outcomes are employed.}




\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}







