\name{bprobgHsCont}
\alias{bprobgHsCont}
\alias{bprobgHsCont3}
\alias{bCopulaCLMgHsCont}
\alias{bprobgHsCont3binTW}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and observed information matrix for 
penalized/unpenalized maximum likelihood optimization when copula models with binary and continuous margins are employed.}




\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}






