\name{bprobgHsContSS}
\alias{bprobgHsContSS}
\alias{bprobgHsCont3SS}
\alias{bprobgHsCont3binTWSS}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and observed information matrix for 
penalized/unpenalized maximum likelihood optimization when copula sample selection models with continuous margins are employed.}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}






