\name{bprobgHsContUniv}
\alias{bprobgHsContUniv}
\alias{bprobgHsContUniv3}
\alias{bprobgHsContUnivBIN}
\alias{bcontSurvGunivInform}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and observed information matrix for 
penalized/unpenalized maximum likelihood optimization when fitting univariate models with discrete/continuous response.}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}






