\name{bprobgHsDiscr1}
\alias{bprobgHsDiscr1}
\alias{bprobgHsDiscr2}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and observed information matrix for 
penalized/unpenalized maximum likelihood optimization when copula models with binary and discrete margins are employed.}




\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}






