\name{bprobgHsDiscr1SS}
\alias{bprobgHsDiscr1SS}
\alias{bprobgHsDiscr2SS}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and observed information matrix for 
penalized/unpenalized maximum likelihood optimization when copula sample selection models with discrete margins are employed.}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}






