\name{bprobgHsPO}
\alias{bprobgHsPO}
\alias{bprobgHsPO0}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and observed or expected information matrix for 
penalized/unpenalized maximum likelihood optimization when bivariate probit models with partial
observability are employed.  

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}






