\name{copgHs}
\alias{copgHs}
\alias{copgHs2}
\alias{copgHs3}
\alias{copgHsAT}
\alias{copgHsCont}

\title{Internal Function}

\description{This and other similar internal functions evaluate the first and second derivatives with respect to the margins and association 
parameter of several copulae.

}


\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}












