\name{distrHs}
\alias{distrHs}
\alias{distrHsAT}
\alias{distrHsAT1}
\alias{distrHsATDiscr}
\alias{distrHsDiscr}

\title{Internal Function}

\description{This and other similar internal functions evaluate the margins' derivatives needed in the likelihood function for the 
binary, discrete and continuous cases.}


\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}












