\name{Dpens}
\alias{Dpens}

\title{Differentiable penalties}

\description{ 
work in progress, temp function
 
}



\usage{

Dpens(params, type = "lasso", lambda = 1, w.alasso = NULL, 
      gamma = 1, a = 3.7, eps = 1e-08)
          
}



\arguments{ 

\item{params}{coefficients.} 

\item{type}{lasso, alasso or scad.} 

\item{lambda}{smoothing parameter.} 

\item{w.alasso}{for alasso.} 

\item{gamma}{default 1.}

\item{a}{for scad.}

\item{eps}{tolerance.}


}


\details{

work in progress.


}


\value{
The function returns a penalty.
}


