\name{eta.tr}
\alias{eta.tr}
\alias{esp.tr}
\alias{enu.tr}
\alias{teta.tr}
\alias{dof.tr}
\alias{form.eq12}
\alias{ass.dp}
\alias{overall.sv}
\alias{overall.svG}
\alias{gamls.upsv}
\alias{pp}
\alias{susutsn}
\alias{susu}
\alias{pscr}
\alias{pscr0}
\alias{startsn}
\alias{bcorrec}
\alias{bcorrecDiscr}
\alias{Cop1Cop2}
\alias{Reg2Copost}
\alias{SS}
\alias{sim.resp}
\alias{r.resp}
\alias{probmS}
\alias{PosDefCor}
\alias{PDef}
\alias{bcorrecFuncs}
\alias{Xdpred}
\alias{mmf}
\alias{vis.gam2}
\alias{cov.c}
\alias{inform.setup}
\alias{intB}
\alias{mice.impute.copulaSS}
\alias{rIC}
\alias{resp.CLM}
\alias{approx.CLM}
\alias{ggmtrust.path}
\alias{ggm.DerivOPT1}
\alias{ggm.DerivOPT2}
\alias{Dpens2}
\alias{survExcInd}
\alias{pTweed}

\title{Internal Function}

\description{This and other similar internal functions map certain key quantities into a feasible parameter space.}


\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}






