\name{g.tri}
\alias{g.tri}
\alias{g.triSS}
\alias{g.triESS}

\title{Internal Function}

\description{This and other similar internal functions calculate the score for trivariate binary models.}



\author{

Author: Panagiota Filippou

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}








