\name{ggmtrust}
\alias{ggmtrust}

\title{ggmtrust for penalised network}

\description{ 
penalised network, work in progress.
 
}



\usage{

ggmtrust(s, n, lambda = 1, pen = "lasso", params = NULL)
          
}



\arguments{ 

\item{s}{Sample covariance matrix.} 

\item{n}{Sample size.} 

\item{lambda}{Regularisation parameter.} 

\item{pen}{Either "lasso" or "ridge".}

\item{params}{If different from null then these are taken as the starting values.}


}


\details{

penalised network, work in progress.


}


\value{
The function returns an object of class \code{ggmtrust}.
}


