\name{H.tri}
\alias{H.tri}
\alias{H.triSS}
\alias{H.triESS}

\title{Internal Function}

\description{This and other similar internal functions calculate the Hessian for trivariate binary models.

}



\author{

Author: Panagiota Filippou

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}








