\name{print.AT}
\alias{print.AT}

\title{Print an AT object}

\description{The print method for an \code{AT} object.}

\usage{


\method{print}{AT}(x, ...)


}


\arguments{ 


\item{x}{\code{AT} object produced by \code{AT()}.}

\item{...}{Other arguments.}

}


\details{ 

\code{print.AT} prints the lower confidence interval limit, estimated AT and upper confidence interval limit. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}


\seealso{
\code{\link{AT}}
}

\keyword{generalised joint regression modelling} 









