\name{print.AT2}
\alias{print.AT2}

\title{Print an AT2 object}

\description{The print method for an \code{AT2} object.}

\usage{


\method{print}{AT2}(x, ...)


}


\arguments{ 


\item{x}{\code{AT2} object produced by \code{AT2()}.}

\item{...}{Other arguments.}

}


\details{ 

\code{print.AT2} prints the lower confidence interval limit, estimated AT and upper confidence interval limit. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}


\seealso{
\code{\link{AT2}}
}

\keyword{generalised joint regression modelling} 









