\name{print.ATE}
\alias{print.ATE}

\title{Print an ATE object}

\description{The print method for an \code{ATE} object.}

\usage{


\method{print}{ATE}(x, ...)


}


\arguments{ 


\item{x}{\code{ATE} object produced by \code{ATE()}.}

\item{...}{Other arguments.}

}


\details{ 

\code{print.ATE} prints the lower confidence interval limit, estimated ATE and upper confidence interval limit. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}


\seealso{
\code{\link{ATE}}
}

\keyword{generalised joint regression modelling} 









