\name{print.cond.mv}
\alias{print.cond.mv}

\title{Print a cond.mv object}

\description{The print method for a \code{cond.mv} object.}

\usage{


\method{print}{cond.mv}(x, ...)


}


\arguments{ 


\item{x}{\code{cond.mv} object produced by \code{cond.mv()}.}

\item{...}{Other arguments.}

}


\details{ 

\code{print.cond.mv} prints the lower confidence interval limit, estimated conditonal mean or variance and upper confidence interval limit. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}


\seealso{
\code{\link{cond.mv}}
}

\keyword{generalised joint regression modelling} 









