\name{print.copulaReg}
\alias{print.copulaReg}

\title{Print a copulaReg object}

\description{The print method for a \code{copulaReg} object.}

\usage{


\method{print}{copulaReg}(x, ...)


}


\arguments{ 


\item{x}{\code{copulaReg} object produced by \code{copulaReg()}.}

\item{...}{Other arguments.}

}


\details{ 

\code{print.copulaReg} prints out the family, model equations, total number of observations, estimated association 
coefficient, etc for the penalized or unpenalized model. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}



\seealso{
\code{\link{copulaReg}}
}


\keyword{generalised joint regression modelling} 








