\name{print.copulaSampleSel}
\alias{print.copulaSampleSel}

\title{Print a copulaSampleSel object}

\description{The print method for a \code{copulaSampleSel} object.}

\usage{


\method{print}{copulaSampleSel}(x, ...)


}


\arguments{ 


\item{x}{\code{copulaSampleSel} object.}

\item{...}{Other arguments.}

}


\details{ 

It prints out the family, model equations, total number of observations, estimated association 
coefficient, etc for the penalized or unpenalized model. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}




\keyword{generalised joint regression modelling} 








