\name{print.gamlss}
\alias{print.gamlss}

\title{Print a gamlss object}

\description{The print method for a \code{gamlss} object.}

\usage{


\method{print}{gamlss}(x, ...)


}


\arguments{ 


\item{x}{\code{gamlss} object produced by \code{gamlss()}.}

\item{...}{Other arguments.}

}


\details{ 

\code{print.gamlss} prints out the family, model equations, total number of observations, etc for the penalized or unpenalized model. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}



\seealso{
\code{\link{gamlss}}
}


\keyword{regression modelling} 








