\name{print.gjrm}
\alias{print.gjrm}

\title{Print a gjrm object}

\description{The print method for a \code{gjrm} object.}

\usage{


\method{print}{gjrm}(x, ...)


}


\arguments{ 


\item{x}{\code{gjrm} object produced by \code{gjrm()}.}

\item{...}{Other arguments.}

}


\details{ 

\code{print.gjrm} prints out the family, model equations, total number of observations, estimated association 
coefficient, etc for the penalized or unpenalized model. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}



\seealso{
\code{\link{gjrm}}
}


\keyword{generalised joint regression modelling} 








