\name{print.marg.mv}
\alias{print.marg.mv}

\title{Print a marg.mv object}

\description{The print method for a \code{marg.mv} object.}

\usage{


\method{print}{marg.mv}(x, ...)


}


\arguments{ 


\item{x}{\code{marg.mv} object produced by \code{marg.mv()}.}

\item{...}{Other arguments.}

}


\details{ 

\code{print.marg.mv} prints the lower confidence interval limit, estimated conditonal mean or variance and upper confidence interval limit. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}


\seealso{
\code{\link{marg.mv}}
}

\keyword{generalised joint regression modelling} 









