\name{print.mb}
\alias{print.mb}

\title{Print an mb object}

\description{The print method for an \code{mb} object.}

\usage{


\method{print}{mb}(x, ...)


}


\arguments{ 


\item{x}{\code{mb} object produced by \code{mb()}.}

\item{...}{Other arguments.}

}


\details{ 

\code{print.mb} prints the lower and upper bounds, confidence interval, and effect assuming random assignment. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}


\seealso{
\code{\link{mb}}
}

\keyword{Manski's bounds} 









