\name{print.OR}
\alias{print.OR}

\title{Print an OR object}

\description{The print method for an \code{OR} object.}

\usage{


\method{print}{OR}(x, ...)


}


\arguments{ 


\item{x}{\code{OR} object produced by \code{OR()}.}

\item{...}{Other arguments.}

}


\details{ 

\code{print.OR} prints the lower confidence interval limit, estimated OR and upper confidence interval limit. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}


\seealso{
\code{\link{OR}}
}

\keyword{generalised joint regression modelling} 









