\name{print.PE}
\alias{print.PE}

\title{Print an PE object}

\description{The print method for an \code{PE} object.}

\usage{


\method{print}{PE}(x, ...)


}


\arguments{ 


\item{x}{\code{PE} object produced by \code{PE()}.}

\item{...}{Other arguments.}

}


\details{ 

\code{print.PE} prints the lower confidence interval limit, estimated PE and upper confidence interval limit. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}


\seealso{
\code{\link{PE}}
}

\keyword{generalised joint regression modelling} 









