\name{print.prev}
\alias{print.prev}

\title{Print an prev object}

\description{The print method for an \code{prev} object.}

\usage{


\method{print}{prev}(x, ...)


}


\arguments{ 


\item{x}{\code{prev} object produced by \code{prev()}.}

\item{...}{Other arguments.}

}


\details{ 

\code{print.prev} prints the lower interval limit, estimated prevalence and upper interval limit. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}




\seealso{
\code{\link{prev}}
}

\keyword{generalised joint regression modelling} 









