\name{print.RR}
\alias{print.RR}

\title{Print an RR object}

\description{The print method for an \code{RR} object.}

\usage{


\method{print}{RR}(x, ...)


}


\arguments{ 


\item{x}{\code{RR} object produced by \code{RR()}.}

\item{...}{Other arguments.}

}


\details{ 

\code{print.RR} prints the lower confidence interval limit, estimated RR and upper confidence interval limit. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}


\seealso{
\code{\link{RR}}
}

\keyword{generalised joint regression modelling} 









