\name{print.SATE}
\alias{print.SATE}

\title{Print an SATE object}

\description{The print method for a \code{SATE} object.}

\usage{


\method{print}{SATE}(x, ...)


}


\arguments{ 


\item{x}{\code{SATE} object produced by \code{SATE()}.}

\item{...}{Other arguments.}

}


\details{ 

\code{print.SATE} prints the lower confidence interval limit, estimated SATE and upper confidence interval limit. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}


\seealso{
\code{\link{SATE}}
}

\keyword{generalised joint regression modelling} 









