\name{print.SemiParBIV}
\alias{print.SemiParBIV}

\title{Print a SemiParBIV object}

\description{The print method for a \code{SemiParBIV} object.}

\usage{


\method{print}{SemiParBIV}(x, ...)


}


\arguments{ 


\item{x}{\code{SemiParBIV} object produced by \code{SemiParBIV()}.}

\item{...}{Other arguments.}

}


\details{ 

\code{print.SemiParBIV} prints out the family, model equations, total number of observations, estimated association 
coefficient and total effective degrees of freedom for the penalized or unpenalized model. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}



\seealso{
\code{\link{SemiParBIV}}
}


\keyword{generalised joint regression modelling} 








