\name{print.SemiParROY}
\alias{print.SemiParROY}

\title{Print a SemiParROY object}

\description{The print method for a \code{SemiParROY} object.}

\usage{


\method{print}{SemiParROY}(x, ...)


}


\arguments{ 


\item{x}{\code{SemiParROY} object.}

\item{...}{Other arguments.}

}


\details{ 

It prints out the family, model equations, total number of observations, estimated association 
coefficient, etc for the penalized or unpenalized model. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}




\keyword{generalised joint regression modelling} 








