\name{print.SemiParTRIV}
\alias{print.SemiParTRIV}

\title{Print a SemiParTRIV object}

\description{The print method for a \code{SemiParTRIV} object.}

\usage{


\method{print}{SemiParTRIV}(x, ...)


}


\arguments{ 


\item{x}{\code{SemiParTRIV} object produced by \code{SemiParTRIV()}.}

\item{...}{Other arguments.}

}


\details{ 

\code{print.SemiParTRIV} prints out the family, model equations, total number of observations, estimated association 
coefficient and total effective degrees of freedom for the penalized or unpenalized model. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}



\seealso{
\code{\link{SemiParTRIV}}
}


\keyword{generalised joint regression modelling} 








