\name{regH}
\alias{regH}

\title{Internal Function}

\description{It applies one of two regularisations on the information matrix if desired. 
These are based on the Cholesky and eigen decompositions. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}







