\name{SemiParBIV.fit.post}
\alias{SemiParBIV.fit.post}
\alias{copulaReg.fit.post}
\alias{copulaSampleSel.fit.post}
\alias{SemiParTRIV.fit.post}
\alias{SemiParROY.fit.post}
\alias{gamlss.fit.post}
\alias{postVb}
\alias{ass.ms}
\alias{edf.loop}
\alias{form.check}
\alias{pred.var} 

\title{Internal Function}

\description{This and other similar internal functions calculate useful post estimation quantities.}


\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}






