\name{triprobgHs}
\alias{triprobgHs}
\alias{triprobgHsSS}
\alias{triprobgHsESS}

\title{Internal Function}

\description{It provides score and Hessian for trivariate binary models.
}



\author{

Author: Panagiota Filippou

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}








