\name{working.comp}
\alias{working.comp}

\title{Internal Function}

\description{It efficiently calculates the working model quantities needed to implement the automatic multiple smoothing parameter estimation 
procedure by exploiting a result which leads to very fast and stable calculations.}




\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}


