\name{bprobgHsContUniv}
\alias{bprobgHsContUniv}
\alias{bprobgHsContUniv3}
\alias{bprobgHsContUnivBIN}
\alias{bcontSurvGunivInform}
\alias{bcontSurvGunivI}
\alias{bcontSurvGunivL}
\alias{bcontSurvGunivMIXED}
\alias{bcontSurvGunivMIXED_ExcessHazard}
\alias{bcontSurvGunivL_ExcessHazard}
\alias{bcontSurvGuniv_ExcessHazard}
\alias{bcontSurvGunivI_ExcessHazard}
\alias{bcontSurvGunivMIXED_LeftTruncation}
\alias{bcontSurvGunivMIXED_ExcessHazard_LeftTruncation}
\alias{bcontSurvG_extended}
\alias{bcontSurvGBINROY}
\alias{bcontSurvGBINss}



\title{Internal Function}

\description{It provides the log-likelihood, gradient and observed information matrix for 
penalized/unpenalized maximum likelihood optimization when fitting univariate models with discrete/continuous response.}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}






