\name{conv.check}
\alias{conv.check}

\title{Some convergence diagnostics}

\description{It takes a fitted model object and produces some 
diagnostic information about the fitting procedure.

}

\usage{

conv.check(x)


}


\arguments{ 


\item{x}{\code{gjrm} object.}


}




\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}



\seealso{
\code{\link{gamlss}}, \code{\link{gjrm}}
}


\keyword{generalised joint regression modelling} 
\keyword{diagnostics} 







